Separate the webhook runner and the action logic
