+++ routemaster/tests/test_layering.py
@@ -45,15 +45,12 @@ DEPENDENCIES = (
     ('state_machine', 'config'),
     ('state_machine', 'feeds'),
     ('state_machine', 'context'),
+    ('state_machine', 'webhooks'),
 
     ('record_states', 'app'),
     ('record_states', 'config'),
 
     ('feeds', 'config'),
-
-    ('actions', 'app'),
-    ('actions', 'config'),
-    ('actions', 'state_machine'),
 )
 
 EXCLUDED_MODULES = (
